Refactor caching logic to AbstractCacheProvider
