+++ Stormpath.SDK/Stormpath.SDK.sln
@@ -19,6 +19,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stormpath.SDK.RestSharpClie
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stormpath.SDK.RestSharpClient.Tests", "Stormpath.SDK.RestSharpClient.Tests\Stormpath.SDK.RestSharpClient.Tests.csproj", "{64893BF7-F4AA-4374-B81B-4572989852EB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stormpath.SDK.Redis", "Stormpath.SDK.Redis\Stormpath.SDK.Redis.csproj", "{5A365FE5-AF6A-4662-A899-071E066146DB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,6 +59,10 @@ Global
 		{64893BF7-F4AA-4374-B81B-4572989852EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{64893BF7-F4AA-4374-B81B-4572989852EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{64893BF7-F4AA-4374-B81B-4572989852EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A365FE5-AF6A-4662-A899-071E066146DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A365FE5-AF6A-4662-A899-071E066146DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A365FE5-AF6A-4662-A899-071E066146DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A365FE5-AF6A-4662-A899-071E066146DB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
