+++ Stormpath.SDK/Stormpath.SDK/Impl/Cache/ICacheResolver.cs
@@ -16,8 +16,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
 using Stormpath.SDK.Cache;
 
 namespace Stormpath.SDK.Impl.Cache
@@ -28,8 +26,8 @@ namespace Stormpath.SDK.Impl.Cache
 
         bool IsAsynchronousSupported { get; }
 
-        ISynchronousCache<string, IDictionary<string, object>> GetCache(Type resourceType);
+        ISynchronousCache<string, IDictionary<string, object>> GetSyncCache(Type resourceType);
 
-        Task<IAsynchronousCache<string, IDictionary<string, object>>> GetCacheAsync(Type resourceType, CancellationToken cancellationToken);
+        IAsynchronousCache<string, IDictionary<string, object>> GetAsyncCache(Type resourceType);
     }
 }
